main advanced by eight commits while this branch was in review, several of
which are the same class of fix landed here in a different form: SteamID by
value (5d3d902), clang -Wformat-nonliteral in notify::FormatV (e769967), a
D3DX9 NuGet CI restore (6a3a3a9), MSVC min/max macro clash (146839a), the
verified-command "actually erase" fix (46a7f48), an x64 netvar refresh with
crash-free pointer resolution (7da2a49), a crash reporter (9408ac7), and a
README rewrite (430dade).

Conflicts:

* Memory.h: main's MemIsReadable-based check is strictly stronger than the
  null/index guards here; taken.  hookData::dst renamed to ::original in this
  branch stays.
* dllmain.cpp: both helpers survive.  ResolveScan (this branch) keeps
  findPattern() nulls out of the pointer arithmetic that dereferences them;
  GuardedVMTHook (main) validates the vtable slot and logs the outcome by
  name.  They do different jobs and are now both used in Main().
* AutoWall.h, Prediction.h, ConVarSpoofing.h, Utils.h, ScriptDumper.h: taken
  this branch's version -- the equivalent fixes on main are narrower.
  Prediction still gates on Settings::Misc::edgeJump and now tracks an
  explicit `active` flag; ConVarSpoofing keeps the one-byte rename and the
  proper Restore() path; GetSteamID returns std::string via
  strutil::FromBounded (safer than nul-terminating info.guid in place).
* README.md, REVIEW.md: fused -- main's netvar-refresh + crash-free init
  section joins the second-pass section here.

MSBuild.  This is the first time the Windows job runs (REVIEW.md called that
out explicitly).  Three defects surfaced:

* Solution config Debug|x86 in the .sln did not match Debug|Win32 in the
  project, so `msbuild /p:Platform=Win32` failed with MSB4126 before touching
  a source file.  The .sln now exposes Win32 at the solution level too.
* drawing.h's `#include <d3dtypes.h>` (the pre-DX8 header) is refused by MSVC
  when d3d9.h is in scope.  D3DTLVERTEX -- the one thing DrawFilledRect needs
  from that world -- is defined locally instead.
* FrameStageNotify.h compared ConVar::intValue (int32_t) with bool settings,
  triggering C4805; static_cast<int> around the bool matches what SetValue
  would have stored anyway.

The min/max regressions the first Windows job hit in AngleMath.h /
notify_core.h / imgui_notify.h had already been fixed on main via
`std::clamp` + parenthesised `(std::min)`; the merge carries them here.

Portable tests: 67 pass under gcc and clang with warnings as errors, and
under ASan + UBSan.  MSBuild remains untested in this environment.

+++ GMod-SDK/hooks/FrameStageNotify.h
@@ -22,7 +22,10 @@ ClientFrameStage_t stage)
 	localPlayer = (C_BasePlayer*)ClientEntityList->GetClientEntity(EngineClient->GetLocalPlayer());
 
 	static ConVar* fullbrightCvar = CVar->FindVar("mat_fullbright");
-	if (fullbrightCvar && fullbrightCvar->intValue != Settings::Visuals::fullBright) {
+	// static_cast around the bool: intValue is int32_t, and comparing them
+	// directly triggers C4805 (unsafe mix of int and bool).  The value 0/1 is
+	// exactly what SetValue below will store anyway.
+	if (fullbrightCvar && fullbrightCvar->intValue != static_cast<int>(Settings::Visuals::fullBright)) {
 		if (Settings::Visuals::fullBright)
 			fullbrightCvar->RemoveFlags(FCVAR_CHEAT);
 		else fullbrightCvar->AddFlags(FCVAR_CHEAT);
@@ -38,10 +41,12 @@ ClientFrameStage_t stage)
 	if (Settings::Misc::svCheats)
 		ConVarSpoofing::Acquire(ConVarSpoofing::cheats, "sv_cheats");
 
-	if (ConVarSpoofing::allowCsLua && ConVarSpoofing::allowCsLua->m_pOriginalCVar->intValue != Settings::Misc::svAllowCsLua)
+	// static_cast<int>: intValue is int32_t and the settings are bool -- see
+	// the note above the fullbright check.
+	if (ConVarSpoofing::allowCsLua && ConVarSpoofing::allowCsLua->m_pOriginalCVar->intValue != static_cast<int>(Settings::Misc::svAllowCsLua))
 		ConVarSpoofing::allowCsLua->m_pOriginalCVar->SetValue(Settings::Misc::svAllowCsLua);
 
-	if (ConVarSpoofing::cheats && ConVarSpoofing::cheats->m_pOriginalCVar->intValue != Settings::Misc::svCheats)
+	if (ConVarSpoofing::cheats && ConVarSpoofing::cheats->m_pOriginalCVar->intValue != static_cast<int>(Settings::Misc::svCheats))
 		ConVarSpoofing::cheats->m_pOriginalCVar->SetValue(Settings::Misc::svCheats);
 
 	//Input->cameraoffset
